WIP refactoring/updating of score_wine() and related methods. New files, wine.py and wine_test.py now contail wine scoring functions and tests respectively. Macklog updated with new issues and newly complketed issues.
